requirements.txt
django~=4.1
djangorestframework~=3.15.1